+++ src/app/layout/header/system-notifications/system-notifications.component.html
@@ -12,7 +12,7 @@
           </div>
           <div class="media-body d-flex flex-row justify-content-between align-items-center">
             <p class="m-0 pr-2 color-white">Accept the transfer of the Asset</p>
-            <span *ngIf="assetsOfferedToMeCount" class="badge badge-danger ml-2">
+            <span *ngIf="assetsOfferedToMeCount" class="badge badge-danger ml-2" data-qa-id='badge-accept-transfer'>
               {{ assetsOfferedToMeCount }}
             </span>
           </div>
@@ -27,7 +27,7 @@
           </div>
           <div class="media-body d-flex flex-row justify-content-between align-items-center">
             <p class="m-0 pr-2 color-white">For approval</p>
-            <span *ngIf="pendingApprovalCount" class="badge badge-danger ml-2">
+            <span *ngIf="pendingApprovalCount" class="badge badge-danger ml-2" data-qa-id='badge-for-approval'>
               {{ pendingApprovalCount }}
             </span>
           </div>
@@ -42,7 +42,7 @@
           </div>
           <div class="media-body d-flex flex-row justify-content-between align-items-center">
             <p class="m-0 pr-2 color-white">Pending Sync to DID Document</p>
-            <span *ngIf="pendingAssetSyncCount" class="badge badge-danger ml-2">
+            <span *ngIf="pendingAssetSyncCount" class="badge badge-danger ml-2" data-qa-id='badge-pending-sync'>
               {{ pendingAssetSyncCount }}
             </span>
           </div>
