+++ src/app/shared/services/enrolment-list/message-subscription.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { MessageSubscriptionService } from './message-subscription.service';
+
+describe('MessageSubscriptionService', () => {
+  let service: MessageSubscriptionService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(MessageSubscriptionService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
