+++ src/app/shared/services/enrolment-list/messaging-facade.service.ts
@@ -1,0 +1,9 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class MessagingFacadeService {
+
+  constructor() { }
+}
